exclusions from code coverage

+++ core/src/main/java/com/mygdx/game/actor/PlayerStatusActor.java
@@ -13,8 +13,8 @@ import com.mygdx.game.Player;
 
 public class PlayerStatusActor extends Actor {
     private Texture texture;
-    public static final int WIDTH = 350;
-    public static final int HEIGHT = 270;
+    public static final float WIDTH = 350;
+    public static final float HEIGHT = 270;
     private final BitmapFont font;
     private Player player;
     private Table info;
